extract actix server builder into separate functions until the proper `Box<dyn ...>` type has been figured out
