+++ src/bin/src/main.rs
@@ -1,49 +1,35 @@
 // Copyright 2025 Sebastian Dobe <sebastiandobe@mailbox.org>
 
 use crate::logging::setup_logging;
-use actix_web::rt::System;
-use actix_web::{App, HttpServer, middleware, web};
-use actix_web_prom::PrometheusMetricsBuilder;
+use actix_web::{middleware, web};
 use cryptr::EncKeys;
-use prometheus::Registry;
-use rauthy_common::constants::{
-    BUILD_TIME, RAUTHY_VERSION, SWAGGER_UI_EXTERNAL, SWAGGER_UI_INTERNAL,
-};
-use rauthy_common::utils::UseDummyAddress;
+use rauthy_common::constants::{BUILD_TIME, RAUTHY_VERSION};
 use rauthy_common::{HTTP_CLIENT, is_sqlite, password_hasher};
 use rauthy_handlers::generic::I18N_CONFIG;
-use rauthy_handlers::openapi::ApiDoc;
 use rauthy_handlers::{
     api_keys, auth_providers, blacklist, clients, dev_only, events, fed_cm, generic, groups, html,
     oidc, roles, scopes, sessions, themes, users,
 };
-use rauthy_middlewares::csrf_protection::CsrfProtectionMiddleware;
-use rauthy_middlewares::ip_blacklist::RauthyIpBlacklistMiddleware;
-use rauthy_middlewares::logging::RauthyLoggingMiddleware;
-use rauthy_middlewares::principal::RauthyPrincipalMiddleware;
 use rauthy_models::app_state::AppState;
 use rauthy_models::database::{Cache, DB};
+use rauthy_models::email;
 use rauthy_models::email::EMail;
 use rauthy_models::entity::pictures::UserPicture;
-use rauthy_models::events::event::Event;
 use rauthy_models::events::health_watch::watch_health;
 use rauthy_models::events::listener::EventListener;
 use rauthy_models::events::notifier::EventNotifier;
 use rauthy_models::events::{init_event_vars, ip_blacklist_handler};
-use rauthy_models::{ListenScheme, email};
 use spow::pow::Pow;
+use std::env;
 use std::error::Error;
-use std::net::Ipv4Addr;
-use std::str::FromStr;
 use std::time::Duration;
-use std::{env, thread};
 use tokio::sync::mpsc;
 use tokio::time;
 use tracing::{debug, error, info, warn};
-use utoipa_swagger_ui::SwaggerUi;
 
 mod dummy_data;
 mod logging;
+mod server;
 mod tls;
 mod version_migration;
 
@@ -244,184 +230,20 @@ https://github.com/sebadob/rauthy/releases/tag/v0.27.0
         tokio::spawn(dummy_data::insert_dummy_data(amount));
     }
 
-    actix_main(app_state).await?;
-    DB::hql().shutdown().await.unwrap();
-
-    Ok(())
-}
-
-async fn actix_main(app_state: web::Data<AppState>) -> std::io::Result<()> {
-    let listen_scheme = app_state.listen_scheme.clone();
-    let listen_addr = app_state.listen_addr.clone();
-
-    let mut workers = env::var("HTTP_WORKERS")
-        .as_deref()
-        .unwrap_or("0")
-        .parse::<usize>()
-        .expect("Unable to parse HTTP_WORKERS");
-    if workers == 0 {
-        workers = num_cpus::get();
-    }
-
-    let swagger = SwaggerUi::new("/docs/v1/swagger-ui/{_:.*}")
-        .url("/docs/v1/api-doc/openapi.json", ApiDoc::build(&app_state))
-        .config(
-            utoipa_swagger_ui::Config::from("../api-doc/openapi.json")
-                .try_it_out_enabled(false)
-                .supported_submit_methods(["get"])
-                .filter(true),
-        );
-
     let metrics_enable = env::var("METRICS_ENABLE")
         .as_deref()
         .unwrap_or("false")
         .parse::<bool>()
         .expect("Cannot parse METRICS_ENABLE to bool");
-    let pub_metrics = if metrics_enable {
-        let shared_registry = Registry::new();
-        let metrics = PrometheusMetricsBuilder::new("api")
-            .registry(shared_registry.clone())
-            .endpoint("/metrics")
-            .exclude("/favicon.ico")
-            .exclude("/metrics")
-            .build()
-            .unwrap();
-
-        let swagger_clone = swagger.clone();
-        thread::spawn(move || {
-            let addr = env::var("METRICS_ADDR").unwrap_or_else(|_| "0.0.0.0".to_string());
-            let port = env::var("METRICS_PORT").unwrap_or_else(|_| "9090".to_string());
-            if let Err(err) = Ipv4Addr::from_str(&addr) {
-                let msg = format!("Error parsing METRICS_ADDR: {}", err);
-                error!(msg);
-                panic!("{}", msg);
-            }
-            let addr_full = format!("{}:{}", addr, port);
-
-            info!("Metrics available on: http://{}/metrics", addr_full);
-            let srv = if *SWAGGER_UI_INTERNAL {
-                info!(
-                    "Serving Swagger UI internally on: http://{}/docs/v1/swagger-ui/",
-                    addr_full
-                );
-                HttpServer::new(move || {
-                    App::new()
-                        .wrap(metrics.clone())
-                        .service(swagger_clone.clone())
-                })
-                .workers(1)
-                .bind(addr_full)
-                .unwrap()
-                .run()
-            } else {
-                HttpServer::new(move || App::new().wrap(metrics.clone()))
-                    .workers(1)
-                    .bind(addr_full)
-                    .unwrap()
-                    .run()
-            };
-            System::new().block_on(srv).unwrap();
-        });
-
-        PrometheusMetricsBuilder::new("rauthy")
-            .registry(shared_registry)
-            // no endpoint means it will not expose one, only collect data
-            .exclude("/favicon.ico")
-            .exclude("/metrics")
-            .build()
-            .unwrap()
+    if metrics_enable {
+        server::server_with_metrics(app_state).await?;
     } else {
-        PrometheusMetricsBuilder::new("rauthy")
-            // no endpoint means it will not expose one, only collect data
-            .exclude_regex(".*")
-            .build()
-            .unwrap()
-    };
-
-    app_state
-        .tx_events
-        .send_async(Event::rauthy_started())
-        .await
-        .unwrap();
-
-    // Note: all .wrap's are executed in reverse order -> the last .wrap is executed as the first
-    // one for any new request
-    let server = HttpServer::new(move || {
-        let mut app = App::new()
-            .wrap(RauthyLoggingMiddleware)
-            .app_data(app_state.clone())
-            .wrap(RauthyPrincipalMiddleware)
-            .wrap(CsrfProtectionMiddleware)
-            .wrap(default_headers())
-            // TODO currently, there is no nice way to conditionally add the public metrics
-            // collecting, apart from specifying the whole server builder twice at this point.
-            // This is due to the very strict typing. Maybe we can extract most builders into
-            // separate functions somehow to get rid of it with disabled metrics.
-            //
-            // It is most probably possible to define `app` with some form of `Box<dyn ...>` to be
-            // able to build 2 different versions of `app` at this point, but I have not figured
-            // out the correct type yet.
-            .wrap(pub_metrics.clone())
-            .service(oidc::get_well_known)
-            .service(fed_cm::get_fed_cm_well_known)
-            .service(generic::catch_all)
-            // Important: Do not move this middleware do need the least amount of computing
-            // for blacklisted IPs -> middlewares are executed in reverse order -> this one first
-            .wrap(RauthyIpBlacklistMiddleware)
-            .service(api_services());
-
-        if *SWAGGER_UI_EXTERNAL {
-            app = app.service(swagger.clone());
-        }
-
-        #[cfg(not(target_os = "windows"))]
-        if matches!(
-            app_state.listen_scheme,
-            ListenScheme::UnixHttp | ListenScheme::UnixHttps
-        ) {
-            app = app.app_data(UseDummyAddress);
-        }
-
-        app
-    })
-    // overwrites the number of worker threads -> default == available cpu cores
-    .workers(workers)
-    .shutdown_timeout(10);
-
-    match listen_scheme {
-        ListenScheme::Http => {
-            server
-                .bind(format!("{}:{}", &listen_addr, get_http_port()))?
-                .run()
-                .await
-        }
-
-        ListenScheme::Https => {
-            server
-                .bind_rustls_0_23(
-                    format!("{}:{}", &listen_addr, get_https_port()),
-                    tls::load_tls().await,
-                )?
-                .run()
-                .await
-        }
-
-        ListenScheme::HttpHttps => {
-            server
-                .bind(format!("{}:{}", &listen_addr, get_http_port()))?
-                .bind_rustls_0_23(
-                    format!("{}:{}", &listen_addr, get_https_port()),
-                    tls::load_tls().await,
-                )?
-                .run()
-                .await
-        }
-
-        #[cfg(not(target_os = "windows"))]
-        ListenScheme::UnixHttp | ListenScheme::UnixHttps => {
-            server.bind_uds(listen_addr)?.run().await
-        }
+        server::server_without_metrics(app_state).await?;
     }
+
+    DB::hql().shutdown().await.unwrap();
+
+    Ok(())
 }
 
 fn get_http_port() -> String {
